+++ my-next-app/services/subjectService.ts
@@ -2,6 +2,7 @@ import { prisma } from "../prisma/prisma";
 import type { NewSubject } from "@/types/subject";
 import type { Subject } from "@/generated/prisma/client";
 import { validateUserSession } from "./authService";
+import { ActionResult } from "@/types/actionResult";
 
 export async function getSubjects(): Promise<Subject[]> {
   try {
@@ -17,32 +18,45 @@ export async function getSubjects(): Promise<Subject[]> {
 export async function postSubject(
   subject: NewSubject,
   sessionId?: string
-): Promise<Subject> {
-  await validateUserSession(sessionId);
+): Promise<ActionResult<Subject>> {
+  const sessionResult = await validateUserSession(sessionId);
+  if (!sessionResult.ok) {
+    return sessionResult;
+  }
+
   try {
-    return await prisma.subject.create({
+    const result = await prisma.subject.create({
       data: {
         name: subject.name,
         monthlyCost: subject.monthlyCost,
       },
     });
+
+    return { ok: true, data: result };
   } catch (error) {
-    console.error("Error creating subject:", error);
-    throw new Error("Error creating subject, please try again");
+    return {
+      ok: false,
+      error: "Error creating subject, please try again",
+    };
   }
 }
 
 export async function deleteSubject(
   subjectId: number,
   sessionId?: string
-): Promise<Subject> {
-  await validateUserSession(sessionId);
+): Promise<ActionResult<Subject>> {
+  const sessionResult = await validateUserSession(sessionId);
+  if (!sessionResult.ok) {
+    return sessionResult;
+  }
+
   try {
-    return await prisma.subject.delete({
+    const subject = await prisma.subject.delete({
       where: { id: subjectId },
     });
+
+    return { ok: true, data: subject };
   } catch (error) {
-    console.error("Error deleting subject:", error);
-    throw new Error("Error deleting subject, please try again");
+    return { ok: false, error: "Error deleting subject, please try again" };
   }
 }
